trigger/requirements.txt
flask==3.*
google-cloud-run==0.10.*
gunicorn==21.*